Update interface for Create workflows
* Attribute and Collection updated
* Entity experiencing interface overflow
* Updated Attribute to AttributeCard
* Updated AttributeCard to AttributePreview
