+++ client/src/components/Navigation/index.tsx
@@ -163,7 +163,7 @@ const Navigation = () => {
             w={"100%"}
             justifyContent={"left"}
             variant={
-              _.startsWith(location.pathname, "/collections")
+              _.includes(location.pathname, "/collection")
                 ? "solid"
                 : "ghost"
             }
@@ -181,7 +181,7 @@ const Navigation = () => {
             w={"100%"}
             justifyContent={"left"}
             variant={
-              _.startsWith(location.pathname, "/entities") ? "solid" : "ghost"
+              _.includes(location.pathname, "/entit") ? "solid" : "ghost"
             }
             onClick={() => navigate("/entities")}
           >
@@ -196,7 +196,7 @@ const Navigation = () => {
             w={"100%"}
             justifyContent={"left"}
             variant={
-              _.startsWith(location.pathname, "/attributes") ? "solid" : "ghost"
+              _.includes(location.pathname, "/attribute") ? "solid" : "ghost"
             }
             onClick={() => navigate("/attributes")}
           >
@@ -210,7 +210,7 @@ const Navigation = () => {
             w={"100%"}
             justifyContent={"left"}
             variant={
-              _.startsWith(location.pathname, "/search") ? "solid" : "ghost"
+              _.includes(location.pathname, "/search") ? "solid" : "ghost"
             }
             leftIcon={<Icon name={"search"} />}
             onClick={() => navigate("/search")}
