+++ client/src/pages/create/Attribute.tsx
@@ -2,7 +2,7 @@
 import React, { useEffect, useState } from "react";
 
 // Existing and custom components
-import { Button, Flex, FormControl, FormErrorMessage, FormLabel, Heading, Input, List, ListItem, Modal, ModalBody, ModalCloseButton, ModalContent, ModalHeader, ModalOverlay, Text, Textarea, useDisclosure } from "@chakra-ui/react";
+import { Button, Flex, FormControl, FormErrorMessage, FormHelperText, FormLabel, Heading, Input, List, ListItem, Modal, ModalBody, ModalCloseButton, ModalContent, ModalHeader, ModalOverlay, Spacer, Text, Textarea, useDisclosure } from "@chakra-ui/react";
 import { Content } from "@components/Container";
 import Icon from "@components/Icon";
 import Values from "@components/Values";
@@ -55,163 +55,141 @@ const Attribute = () => {
   };
 
   return (
-    <Content>
-      <Flex direction={"column"} p={"2"} gap={"6"} wrap={"wrap"}>
-        <Flex direction={"column"} w={"100%"} p={"4"} bg={"white"}>
-          {/* Page header */}
-          <Flex direction={"column"} p={"2"} pt={"4"} pb={"4"}>
-            <Flex direction={"row"} align={"center"} justify={"space-between"}>
-              <Heading fontWeight={"semibold"}>Create Attribute</Heading>
-              <Button
-                rightIcon={<Icon name={"info"} />}
-                variant={"outline"}
-                onClick={onOpen}
-              >
-                Info
-              </Button>
-            </Flex>
-          </Flex>
-
-          <Flex
-            direction={"column"}
-            justify={"center"}
-            align={"center"}
-            gap={"6"}
-            p={"2"}
-            pb={"6"}
-            mb={["12", "8"]}
+    <Content vertical>
+      <Flex direction={"column"} gap={"6"} w={"100%"} h={"100%"} p={"4"} bg={"white"} maxW={"4xl"}>
+        {/* Page header */}
+        <Flex direction={"row"} align={"center"} justify={"space-between"}>
+          <Heading fontWeight={"semibold"} size={"lg"}>Create a new template Attribute</Heading>
+          <Button
+            rightIcon={<Icon name={"info"} />}
+            variant={"outline"}
+            onClick={onOpen}
           >
-            <Flex direction={"column"} w={"100%"}>
-              <Heading fontWeight={"semibold"} size={"lg"}>
-                Details
-              </Heading>
-              <Text>
-                Specify some basic details about this template Attribute. The
-                metadata associated with this template should be specified using
-                Values.
-              </Text>
-            </Flex>
+            Info
+          </Button>
+        </Flex>
 
-            <Flex
-              direction={"column"}
-              gap={"2"}
-              w={"100%"}
-              maxW={"4xl"}
-              wrap={["wrap", "nowrap"]}
-            >
-              <Flex direction={"row"} gap={"4"} wrap={["wrap", "nowrap"]}>
-                <FormControl isRequired>
-                  <FormLabel>Name</FormLabel>
-                  <Input
-                    placeholder={"Name"}
-                    value={name}
-                    onChange={(event) => setName(event.target.value)}
-                    required
-                  />
-                  {isNameError && (
-                    <FormErrorMessage>
-                      A name must be specified for the Attribute.
-                    </FormErrorMessage>
-                  )}
-                </FormControl>
+        <Flex direction={"column"} gap={"4"}>
+          <FormControl isRequired>
+            <FormLabel>Attribute Name</FormLabel>
+            <Input
+              w={["100%", "md"]}
+              placeholder={"Name"}
+              value={name}
+              onChange={(event) => setName(event.target.value)}
+              required
+            />
+            {isNameError && (
+              <FormErrorMessage>
+                A name must be specified for the Attribute.
+              </FormErrorMessage>
+            )}
+            <FormHelperText>Provide a concise and descriptive name for the template Attribute.</FormHelperText>
+          </FormControl>
 
-                <FormControl isRequired>
-                  <FormLabel>Description</FormLabel>
-                  <Textarea
-                    value={description}
-                    placeholder={"Attribute Description"}
-                    onChange={(event) => setDescription(event.target.value)}
-                  />
-                  {isDescriptionError && (
-                    <FormErrorMessage>
-                      A description should be provided for the Attribute.
-                    </FormErrorMessage>
-                  )}
-                </FormControl>
-              </Flex>
+          <FormControl isRequired>
+            <FormLabel>Attribute Description</FormLabel>
+            <Textarea
+              w={["100%", "lg"]}
+              value={description}
+              placeholder={"Description"}
+              onChange={(event) => setDescription(event.target.value)}
+            />
+            {isDescriptionError && (
+              <FormErrorMessage>
+                A description should be provided for the Attribute.
+              </FormErrorMessage>
+            )}
+            <FormHelperText>Describe the purpose and contents of this template Attribute.</FormHelperText>
+          </FormControl>
+        </Flex>
 
-              <Flex>
-                <FormControl isRequired isInvalid={isValueError}>
-                  <FormLabel>Values</FormLabel>
-                  <Values
-                    collection={values}
-                    viewOnly={false}
-                    setValues={setValues}
-                  />
-                </FormControl>
-              </Flex>
-            </Flex>
-          </Flex>
+        <Flex w={"100%"}>
+          <FormControl isRequired isInvalid={isValueError}>
+            <FormLabel>Values</FormLabel>
+            <Values
+              collection={values}
+              viewOnly={false}
+              setValues={setValues}
+            />
+            <FormHelperText>Add Values to the template Attribute. Values must be named, but they are not required to have data associated with them. Data can be specified if desired.</FormHelperText>
+          </FormControl>
+        </Flex>
 
-          {/* Action buttons */}
-          <Flex
-            direction={"row"}
-            flexWrap={"wrap"}
-            gap={"6"}
-            justify={"space-between"}
-            alignSelf={"center"}
-            w={"100%"}
-            p={"4"}
+        {/* Action buttons */}
+        <Flex
+          direction={"row"}
+          wrap={"wrap"}
+          gap={"6"}
+          p={"4"}
+        >
+          <Button
+            colorScheme={"red"}
+            variant={"outline"}
+            rightIcon={<Icon name={"cross"} />}
+            onClick={() => navigate("/attributes")}
           >
-            <Button
-              colorScheme={"red"}
-              variant={"outline"}
-              rightIcon={<Icon name={"cross"} />}
-              onClick={() => navigate("/attributes")}
-            >
-              Cancel
-            </Button>
-            <Button
-              colorScheme={"green"}
-              rightIcon={<Icon name={"check"} />}
-              onClick={onSubmit}
-              isDisabled={isDetailsError && !isSubmitting}
-            >
-              Finish
-            </Button>
-          </Flex>
+            Cancel
+          </Button>
+          <Spacer />
+          <Button
+            colorScheme={"green"}
+            rightIcon={<Icon name={"check"} />}
+            onClick={onSubmit}
+            isDisabled={isDetailsError && !isSubmitting}
+          >
+            Finish
+          </Button>
         </Flex>
       </Flex>
 
       {/* Information modal */}
-      <Modal isOpen={isOpen} onClose={onClose}>
+      <Modal isOpen={isOpen} onClose={onClose} size={"2xl"}>
         <ModalOverlay />
         <ModalContent>
-          <ModalHeader>Attributes</ModalHeader>
+          <ModalHeader>Creating Template Attributes</ModalHeader>
           <ModalCloseButton />
           <ModalBody>
-            <Flex direction={"column"} gap={"4"} p={"2"}>
-              <Text>There are six supported types of metadata:</Text>
+            <Flex direction={"column"} gap={"4"} p={"4"}>
+              <Heading size={"md"}>Overview</Heading>
+              <Text>
+                Create a new template Attribute to be used to specify metadata associated with Entities.
+                Using Values, predefined metadata fields can be associated with Entities.
+                After creating a template Attribute, it can be used during the Entity creation process to pre-populate Attribute information and Entity metadata.
+              </Text>
+
+              <Heading size={"md"}>Values</Heading>
+              <Text>There are six supported Value types:</Text>
               <List spacing={2}>
                 <ListItem>
                   <Flex gap={"2"} align={"center"}>
                     <Icon name={"v_date"} color={"orange.300"} />
-                    <Text>Date: Used to specify a point in time.</Text>
+                    <Text><b>Date:</b> Used to specify a point in time.</Text>
                   </Flex>
                 </ListItem>
                 <ListItem>
                   <Flex gap={"2"} align={"center"}>
                     <Icon name={"v_text"} color={"blue.300"} />
-                    <Text>Text: Used to specify text of variable length.</Text>
+                    <Text><b>Text:</b> Used to specify text of variable length.</Text>
                   </Flex>
                 </ListItem>
                 <ListItem>
                   <Flex gap={"2"} align={"center"}>
                     <Icon name={"v_number"} color={"green.300"} />
-                    <Text>Number: Used to specify a numerical value.</Text>
+                    <Text><b>Number:</b> Used to specify a numerical value.</Text>
                   </Flex>
                 </ListItem>
                 <ListItem>
                   <Flex gap={"2"} align={"center"}>
                     <Icon name={"v_url"} color={"yellow.300"} />
-                    <Text>URL: Used to specify a link.</Text>
+                    <Text><b>URL:</b> Used to specify a link.</Text>
                   </Flex>
                 </ListItem>
                 <ListItem>
                   <Flex gap={"2"} align={"center"}>
                     <Icon name={"entity"} color={"purple.300"} />
                     <Text>
-                      Entity: Used to specify a relation to another Entity.
+                      <b>Entity:</b> Used to specify a relation to another Entity.
                     </Text>
                   </Flex>
                 </ListItem>
@@ -219,12 +197,12 @@ const Attribute = () => {
                   <Flex gap={"2"} align={"center"}>
                     <Icon name={"v_select"} color={"teal.300"} />
                     <Text>
-                      Select: Used to specify an option from a group of options.
+                      <b>Select:</b> Used to specify an option from a group of options.
                     </Text>
                   </Flex>
                 </ListItem>
               </List>
-              <Text>Values can be specified using the buttons.</Text>
+              <Text>Values can be added to the template Attribute by clicking "Add Value" and selecting the specific type of Value.</Text>
             </Flex>
           </ModalBody>
         </ModalContent>
